requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
opencv-python==4.8.0.76
Pillow==10.0.0
pandas==2.0.3
numpy==1.24.3